+++ examples/fewer/CMakeLists.txt
@@ -14,7 +14,6 @@ project(fewer)
 
 set(EXECUTABLE_OUTPUT_PATH "bin")
 set(LSB_RELEASE "")
-set(ETC_RELEASE "")
 
 if(MSVC)
     set(CMAKE_LIBRARY_OUTPUT_DIRECTORY "${EXECUTABLE_OUTPUT_PATH}")
@@ -28,14 +27,13 @@ if(MSVC)
     set(CMAKE_RUNTIME_OUTPUT_DIRECTORY_RELEASE "${EXECUTABLE_OUTPUT_PATH}")
 elseif(CMAKE_SYSTEM_NAME MATCHES "Linux")
     execute_process(COMMAND lsb_release -a OUTPUT_VARIABLE LSB_RELEASE)
-    execute_process(COMMAND cat /etc/*release* OUTPUT_VARIABLE ETC_RELEASE)
 endif()
 
 set(WINDOWS_C_FLAGS "${CMAKE_C_FLAGS} /O2 /WX /W4 /wd4204")
-
 set(UNIX_C_FLAGS "${CMAKE_C_FLAGS} -O2 -Werror")
 set(GNU_C_FLAGS "${UNIX_C_FLAGS} -std=gnu99 -Wextra -Wall -pedantic")
-set(CLANG_C_FLAGS "${GNU_C_FLAGS} -Wmost -Weverything")
+set(LLVM_FLAGS "-Wmost -Weverything -Wno-documentation-unknown-command")
+set(CLANG_C_FLAGS "${GNU_C_FLAGS} ${LLVM_FLAGS}")
 
 if(MSVC)
     set(CMAKE_C_FLAGS "${WINDOWS_C_FLAGS}")
@@ -50,7 +48,9 @@ else()
     set(CMAKE_C_FLAGS "${UNIX_C_FLAGS}")
 endif()
 
-add_executable(fewer lib/main.c lib/fewer.c)
+include_directories(include)
+
+add_executable(fewer src/cmd/fewer/main.c src/fewer.c)
 
 if("$ENV{CLOUD_ABI_TRIPLET}" MATCHES "cloudabi")
     add_test(NAME fewer_unit_test COMMAND sh -c "cloudabi-run -e bin/fewer <cloudabi-unittest.yml")
@@ -68,9 +68,9 @@ else()
     add_test(NAME fewer_help_test COMMAND bin/fewer -h)
 endif()
 
-add_custom_target(cppcheck COMMAND cppcheck -q --enable=all --inline-suppr --suppressions-list=suppressions.cfg -DX_PATH_MAX=1024 --error-exitcode=1 lib)
-add_custom_target(cpplint COMMAND sail lib | xargs -n 1 cpplint)
-add_custom_target(vera++ COMMAND sail lib | xargs -n 1 vera++)
+add_custom_target(cppcheck COMMAND cppcheck -q --enable=all --inline-suppr --suppressions-list=suppressions.cfg -DX_PATH_MAX=1024 --error-exitcode=1 -I include src)
+add_custom_target(cpplint COMMAND sail include src | xargs -n 1 cpplint)
+add_custom_target(vera++ COMMAND sail include src | xargs -n 1 vera++)
 
 # Missing linter ports
 if(
@@ -137,5 +137,7 @@ if(
 else()
     add_custom_target(leaks DEPENDS valgrind)
 endif()
+
+add_custom_target(doc COMMAND doxygen Doxyfile)
 
 add_custom_target(clean-cmake COMMAND sh clean-cmake.sh)
