+++ examples/fewer/CPPLINT.cfg
@@ -4,9 +4,11 @@ filter=-build/include_subdir
 filter=-legal/copyright
 filter=-readability/braces
 filter=-readability/casting
+filter=-readability/namespace
 filter=-readability/streams
 filter=-runtime/int
 filter=-runtime/arrays
+filter=-whitespace/indent
 filter=-whitespace/line_length
 filter=-whitespace/newline
 filter=-whitespace/parens
