--- examples/fewer/lib/main.h
@@ -1,9 +1,0 @@
-#pragma once
-
-// Copyright 2017 Andrew Pennebaker
-
-#if defined(__SVR4)
-#define _POSIX_C_SOURCE 200809L
-#endif
-
-void usage(char **argv);
