+++ examples/fewer/sample.envrc
@@ -1,0 +1,10 @@
+#!/bin/bash
+#
+# Install:
+# cp sample.envrc .envrc
+#
+# Uninstall:
+# rm .envrc
+
+source_up
+source "$HOME/.direnv.d/pyenv.sh"
